feat: add ProjectDetailsHero, ProjectItem, and ProjectLinks components; refactor caseStudyHelpers and update imports
