+++ src/app/projects/[slug]/components/ProjectDetailsHero.tsx
