+++ src/app/projects/[slug]/components/ProjectItem.tsx
