+++ src/app/projects/[slug]/utils/caseStudyHelpers.ts
